# static/openapi.yaml
openapi: 3.0.3
info:
  title: Archery Results Tracker API
  version: 1.0.0
  description: Track and visualize archery competition results

servers:
  - url: http://localhost:5001

tags:
  - name: archers
  - name: results
  - name: charts
  - name: import
  - name: sync
  - name: backup

paths:

  /api/archers:
    get:
      tags: [archers]
      summary: List all archers
      responses:
        '200':
          description: Array of archer objects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Archer' }
    post:
      tags: [archers]
      summary: Add a new archer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "John Doe" }
                profile_url: { type: string, example: "/Archer/Details/1200" }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  name: { type: string }
        '400':
          description: name is required

  /api/archers/{archer_id}/results:
    get:
      tags: [archers]
      summary: Get results for an archer
      parameters:
        - { name: archer_id, in: path, required: true, schema: { type: integer } }
        - { name: category, in: query, schema: { type: string, example: "C1" } }
        - { name: distance, in: query, schema: { type: string, example: "18 m" } }
        - { name: date_from, in: query, schema: { type: string, format: date, example: "2024-01-01" } }
        - { name: date_to, in: query, schema: { type: string, format: date, example: "2024-12-31" } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Result' }

  /api/archers/{archer_id}/stats:
    get:
      tags: [archers]
      summary: Get statistics for an archer
      parameters:
        - { name: archer_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArcherStats' }

  /api/results:
    get:
      tags: [results]
      summary: Get all results with optional filters
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: distance, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Result' }

  /api/categories:
    get:
      tags: [results]
      summary: List all distinct categories
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code: { type: string, example: "C1" }
                    description: { type: string, example: "Compound Men" }

  /api/distances:
    get:
      tags: [results]
      summary: List all distinct distances
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { type: string, example: "18 m" }

  /api/import:
    post:
      tags: [import]
      summary: Import results from HTML content or server-side file path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [archer_name]
              properties:
                archer_name: { type: string }
                html_content: { type: string, description: "Raw HTML string" }
                html_file: { type: string, description: "Server-side file path" }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportResult' }
        '400':
          description: Missing required fields

  /api/import/file:
    post:
      tags: [import]
      summary: Import results from uploaded HTML file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archer_name, file]
              properties:
                archer_name: { type: string }
                file: { type: string, format: binary }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportResult' }
        '400':
          description: Missing file or archer_name

  /api/chart/scores/{archer_id}:
    get:
      tags: [charts]
      summary: Time-series chart data for a single archer grouped by distance/category
      parameters:
        - { name: archer_id, in: path, required: true, schema: { type: integer } }
        - { name: category, in: query, schema: { type: string } }
        - { name: distance, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChartResponse' }

  /api/chart/compare:
    get:
      tags: [charts]
      summary: Compare multiple archers on the same time-series chart
      parameters:
        - name: archer_ids
          in: query
          required: true
          schema: { type: array, items: { type: integer } }
          style: form
          explode: true
        - { name: category, in: query, schema: { type: string } }
        - { name: distance, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChartResponse' }
        '400':
          description: archer_ids required

  /api/chart/yearly/{archer_id}:
    get:
      tags: [charts]
      summary: Per-year average and best scores for a single archer
      parameters:
        - { name: archer_id, in: path, required: true, schema: { type: integer } }
        - { name: category, in: query, schema: { type: string } }
        - { name: distance, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  archer_name: { type: string }
                  datasets:
                    type: array
                    items:
                      type: object
                      properties:
                        label: { type: string, example: "Snitt" }
                        data: { type: array, items: { type: number } }
                        years: { type: array, items: { type: string } }
                        competitions: { type: array, items: { type: integer } }

  /api/chart/yearly/compare:
    get:
      tags: [charts]
      summary: Compare yearly averages across multiple archers
      parameters:
        - name: archer_ids
          in: query
          required: true
          schema: { type: array, items: { type: integer } }
          style: form
          explode: true
        - { name: category, in: query, schema: { type: string } }
        - { name: distance, in: query, schema: { type: string } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChartResponse' }
        '400':
          description: archer_ids required

  /api/backup:
    post:
      tags: [backup]
      summary: Create a database backup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, example: "before_migration" }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupInfo' }

  /api/backup/list:
    get:
      tags: [backup]
      summary: List all available backups
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BackupInfo' }

  /api/backup/restore:
    post:
      tags: [backup]
      summary: Restore a backup (auto-creates safety snapshot first)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename]
              properties:
                filename: { type: string, example: "archery_20250330_143000.db" }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  restored_from: { type: string }
                  safety_backup: { type: string }
        '400':
          description: filename is required
        '404':
          description: Backup file not found

  /api/sync/status:
    get:
      tags: [sync]
      summary: Overall sync statistics
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_archers_found: { type: integer }
                  total_not_found: { type: integer }
                  historical_synced: { type: integer }
                  total_results: { type: integer }
                  unresolved_errors: { type: integer }
                  last_sync: { type: string, format: date-time, nullable: true }
                  max_external_id: { type: integer }

  /api/sync/archers:
    get:
      tags: [sync]
      summary: Sync status for all known archers
      parameters:
        - { name: only_existing, in: query, schema: { type: boolean, default: true } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '200':
          content:
            application/json:
              schema: { type: array, items: { type: object } }

  /api/sync/logs:
    get:
      tags: [sync]
      summary: Recent sync job logs
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200':
          content:
            application/json:
              schema: { type: array, items: { type: object } }

  /api/sync/errors:
    get:
      tags: [sync]
      summary: Unresolved sync errors
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
      responses:
        '200':
          content:
            application/json:
              schema: { type: array, items: { type: object } }

  /api/sync/errors/{error_id}/resolve:
    post:
      tags: [sync]
      summary: Mark a sync error as resolved
      parameters:
        - { name: error_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }

  /api/sync/start:
    post:
      tags: [sync]
      summary: Start a background sync operation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                job_type:
                  type: string
                  enum: [daily, discovery, historical, single]
                  default: daily
                archer_id: { type: integer, description: "For job_type=single" }
                archer_ids: { type: array, items: { type: integer }, description: "For job_type=historical" }
                start_id: { type: integer, description: "For job_type=discovery" }
                end_id: { type: integer, description: "For job_type=discovery" }
                years: { type: array, items: { type: integer }, description: "For job_type=historical" }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  job_type: { type: string }
        '409':
          description: Sync already running

  /api/sync/stop:
    post:
      tags: [sync]
      summary: Stop the currently running sync
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '404':
          description: No sync running

  /api/sync/running:
    get:
      tags: [sync]
      summary: Check if a sync is currently running
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  running: { type: boolean }

components:
  schemas:
    Archer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        external_id: { type: integer, nullable: true }
        club: { type: string, nullable: true }
        profile_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

    Result:
      type: object
      properties:
        id: { type: integer }
        archer_id: { type: integer }
        archer_name: { type: string }
        event_id: { type: integer }
        event_name: { type: string }
        date: { type: string, format: date }
        distance: { type: string, example: "18 m" }
        category: { type: string, example: "C1" }
        score: { type: integer }
        score_per_60:
          type: number
          description: Score normalised to 60-arrow basis
          example: 540.0
        placement: { type: integer, nullable: true }

    ArcherStats:
      type: object
      properties:
        total_competitions: { type: integer }
        avg_score: { type: number }
        best_score: { type: integer }
        worst_score: { type: integer }
        first_places: { type: integer }
        by_category:
          type: array
          items:
            type: object
            properties:
              category: { type: string }
              competitions: { type: integer }
              avg_score: { type: number }
              best_score: { type: integer }

    ChartResponse:
      type: object
      properties:
        datasets:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              data: { type: array, items: { type: number } }
              score_per_60: { type: array, items: { type: number } }
              dates: { type: array, items: { type: string, format: date } }

    ImportResult:
      type: object
      properties:
        message: { type: string }
        archer_id: { type: integer }
        imported_count: { type: integer }

    BackupInfo:
      type: object
      properties:
        filename: { type: string, example: "archery_20250330_143000.db" }
        path: { type: string }
        size_bytes: { type: integer }
        timestamp: { type: string, example: "20250330_143000" }
        created_at: { type: string, format: date-time }